Create isAnagram
Write a program that finds out whether two words are anagrams or not. If one word is an anagram of another, print "yes", else print "no".

Note: anagrams are words that contain the same characters with the same frequencies. In other words, anagrams are created by changing the order of letters in the word.

For example:

"ppaaagg" (2 ps, 3 as, 2 gs) and "gagaapp" (2 ps, 3 as, 2 gs) are anagrams;
"hello" (1 h, 1 e, 2 ls, 1 o) and "helllo" (1 h, 1 e, 3 ls, 1 o) are not anagrams.
Remember: anagrams are case-insensitive.